Revert "Gpload add case (#11341)"
This reverts commit 6ebd75355fce38f8b9af883e5ae4510164517f08.
