+++ gpMgmt/bin/gpload_test/.gitignore
@@ -4,12 +4,9 @@ gpload/config
 gpload2/config
 gpload2/setup.out
 gpload2/data_file.txt
-gpload2/data_file2.txt
-gpload2/data_file.csv
-gpload2/data/large_file.csv
-gpload2/data_file*
 GPTest.pm
 gpstringsubs.pl
 gpdiff.pl
 atmsort.pm
 explain.pm
+data/large_file.csv
