--- gpMgmt/bin/gpload_test/gpload2/data/external_file_17.txt
@@ -1,7 +1,0 @@
-Line 1;1;2012-06-01 15:30:30;1.2
-2nd line;2;2012-06-01 15:30:40;2.3
-test;3;2012-06-01 15:30:50;3
-Vide;4;2012-06-01 15:30:10;4.5
-Field 2;5;2012-06-01 15:10:30;5
-new line;6;2012-06-01 17:30:30;5.6
-Line 10;7;2012-06-02 15:30:30;6.6
