+++ gpMgmt/bin/gpload_test/gpload2/query44.ans
@@ -1,18 +1,18 @@
-2021-01-08 15:59:10|INFO|gpload session started 2021-01-08 15:59:10
-2021-01-08 15:59:10|INFO|setting schema 'public' for table 'texttable'
-2021-01-08 15:59:10|INFO|started gpfdist -p 8081 -P 8082 -f "pathto/data_file.txt" -t 30
-2021-01-08 15:59:10|INFO|did not find an external table to reuse. creating ext_gpload_reusable_63e1b31e_5187_11eb_b549_00505698707d
-2021-01-08 15:59:10|INFO|running time: 0.10 seconds
-2021-01-08 15:59:10|INFO|rows Inserted          = 16
-2021-01-08 15:59:10|INFO|rows Updated           = 0
-2021-01-08 15:59:10|INFO|data formatting errors = 0
-2021-01-08 15:59:10|INFO|gpload succeeded
-2021-01-08 15:59:10|INFO|gpload session started 2021-01-08 15:59:10
-2021-01-08 15:59:10|INFO|setting schema 'public' for table 'texttable'
-2021-01-08 15:59:10|INFO|started gpfdist -p 8081 -P 8082 -f "pathto/data_file.txt" -t 30
-2021-01-08 15:59:10|INFO|reusing external table ext_gpload_reusable_63e1b31e_5187_11eb_b549_00505698707d
-2021-01-08 15:59:10|INFO|running time: 0.11 seconds
-2021-01-08 15:59:10|INFO|rows Inserted          = 16
-2021-01-08 15:59:10|INFO|rows Updated           = 0
-2021-01-08 15:59:10|INFO|data formatting errors = 0
-2021-01-08 15:59:10|INFO|gpload succeeded
+2020-11-30 16:20:46|INFO|gpload session started 2020-11-30 16:20:46
+2020-11-30 16:20:46|INFO|setting schema 'public' for table 'texttable'
+2020-11-30 16:20:46|INFO|started gpfdist -p 8081 -P 8082 -f "/home/gpadmin/workspace/gpdb/gpMgmt/bin/gpload_test/gpload2/data_file.txt" -t 30
+2020-11-30 16:20:46|INFO|reusing external table ext_gpload_reusable_f160fcb6_32e4_11eb_b17e_00505698707d
+2020-11-30 16:20:46|INFO|running time: 0.12 seconds
+2020-11-30 16:20:46|INFO|rows Inserted          = 16
+2020-11-30 16:20:46|INFO|rows Updated           = 0
+2020-11-30 16:20:46|INFO|data formatting errors = 0
+2020-11-30 16:20:46|INFO|gpload succeeded
+2020-11-30 16:20:47|INFO|gpload session started 2020-11-30 16:20:47
+2020-11-30 16:20:47|INFO|setting schema 'public' for table 'texttable'
+2020-11-30 16:20:47|INFO|started gpfdist -p 8081 -P 8082 -f "/home/gpadmin/workspace/gpdb/gpMgmt/bin/gpload_test/gpload2/data_file.txt" -t 30
+2020-11-30 16:20:47|INFO|reusing external table ext_gpload_reusable_f160fcb6_32e4_11eb_b17e_00505698707d
+2020-11-30 16:20:47|INFO|running time: 0.11 seconds
+2020-11-30 16:20:47|INFO|rows Inserted          = 16
+2020-11-30 16:20:47|INFO|rows Updated           = 0
+2020-11-30 16:20:47|INFO|data formatting errors = 0
+2020-11-30 16:20:47|INFO|gpload succeeded
