--- gpMgmt/bin/gpload_test/gpload2/query73.ans
@@ -1,14 +1,0 @@
-2020-12-14 17:19:12|INFO|gpload session started 2020-12-14 17:19:12
-2020-12-14 17:19:12|INFO|setting schema 'public' for table 'chinese表'
-2020-12-14 17:19:12|ERROR|no mapping for input column "notexist" to output table
-2020-12-14 17:19:12|INFO|rows Inserted          = 0
-2020-12-14 17:19:12|INFO|rows Updated           = 0
-2020-12-14 17:19:12|INFO|data formatting errors = 0
-2020-12-14 17:19:12|INFO|gpload failed
-2020-12-14 17:19:12|INFO|gpload session started 2020-12-14 17:19:12
-2020-12-14 17:19:12|INFO|setting schema 'public' for table 'chinese表'
-2020-12-14 17:19:12|ERROR|no mapping for input column "notexist" to output table
-2020-12-14 17:19:12|INFO|rows Inserted          = 0
-2020-12-14 17:19:12|INFO|rows Updated           = 0
-2020-12-14 17:19:12|INFO|data formatting errors = 0
-2020-12-14 17:19:12|INFO|gpload failed
