--- gpMgmt/bin/gpload_test/gpload2/query76.ans
@@ -1,45 +1,0 @@
-WARNING:  nonstandard use of \\ in a string literal
-LINE 13:                  and pgext.fmtopts like '%delimiter '';'' nu...
-                                                 ^
-HINT:  Use the escape string syntax for backslashes, e.g., E'\\'.
-WARNING:  nonstandard use of escape in a string literal
-LINE 1: .../data_file.txt') format'text' (delimiter ';' null '\N' escap...
-                                                             ^
-HINT:  Use the escape string syntax for escapes, e.g., E'\r\n'.
-WARNING:  nonstandard use of \' in a string literal
-LINE 1: ...xt') format'text' (delimiter ';' null '\N' escape '\') encod...
-                                                             ^
-HINT:  Use '' to write quotes in strings, or use the escape string syntax (E'...').
-2021-01-08 16:58:46|INFO|gpload session started 2021-01-08 16:58:46
-2021-01-08 16:58:46|INFO|setting schema 'public' for table 'chinese表'
-2021-01-08 16:58:46|INFO|started gpfdist -p 8081 -P 8082 -f "/home/gpadmin/workspace/gpdb/gpMgmt/bin/gpload_test/gpload2/data_file.txt" -t 30
-2021-01-08 16:58:46|INFO|did not find an external table to reuse. creating ext_gpload_reusable_b76b0b68_518f_11eb_9406_00505698707d
-2021-01-08 16:58:46|ERROR|unexpected error -- backtrace written to log file
-2021-01-08 16:58:46|INFO|rows Inserted          = 0
-2021-01-08 16:58:46|INFO|rows Updated           = 0
-2021-01-08 16:58:46|INFO|data formatting errors = 0
-2021-01-08 16:58:46|INFO|gpload failed
-2021-01-08 16:58:46|ERROR|configuration file error: expected <block end>, but found '<scalar>', line 14
-2021-01-08 16:58:47|INFO|gpload session started 2021-01-08 16:58:47
-2021-01-08 16:58:47|INFO|setting schema 'public' for table 'chinese表'
-2021-01-08 16:58:47|ERROR|no mapping for input column "'列1'" to output table
-2021-01-08 16:58:47|INFO|rows Inserted          = 0
-2021-01-08 16:58:47|INFO|rows Updated           = 0
-2021-01-08 16:58:47|INFO|data formatting errors = 0
-2021-01-08 16:58:47|INFO|gpload failed
-NOTICE:  Table doesn't have 'DISTRIBUTED BY' clause -- Using column named '列1' as the Greenplum Database data distribution key for this table.
-HINT:  The 'DISTRIBUTED BY' clause determines the distribution of data. Make sure column(s) chosen are the optimal data distribution key to minimize skew.
-WARNING:  nonstandard use of \\ in a string literal
-LINE 13:                  and pgext.fmtopts like '%delimiter '';'' nu...
-                                                 ^
-HINT:  Use the escape string syntax for backslashes, e.g., E'\\'.
-2021-01-08 16:58:47|INFO|gpload session started 2021-01-08 16:58:47
-2021-01-08 16:58:47|INFO|setting schema 'public' for table 'chinese表'
-2021-01-08 16:58:47|INFO|started gpfdist -p 8081 -P 8082 -f "/home/gpadmin/workspace/gpdb/gpMgmt/bin/gpload_test/gpload2/data_file.txt" -t 30
-2021-01-08 16:58:47|INFO|did not find a staging table to reuse. creating staging_gpload_reusable_5171458efa83aaf8c5bc7004bae85d5b
-2021-01-08 16:58:47|INFO|did not find an external table to reuse. creating ext_gpload_reusable_b7f9950e_518f_11eb_93a8_00505698707d
-2021-01-08 16:58:47|ERROR|unexpected error -- backtrace written to log file
-2021-01-08 16:58:47|INFO|rows Inserted          = 0
-2021-01-08 16:58:47|INFO|rows Updated           = 0
-2021-01-08 16:58:47|INFO|data formatting errors = 0
-2021-01-08 16:58:47|INFO|gpload failed
