--- gpMgmt/bin/gpload_test/gpload2/query77.ans
@@ -1,59 +1,0 @@
-WARNING:  nonstandard use of \\ in a string literal
-LINE 13:                  and pgext.fmtopts like '%delimiter '';'' nu...
-                                                 ^
-HINT:  Use the escape string syntax for backslashes, e.g., E'\\'.
-WARNING:  nonstandard use of escape in a string literal
-LINE 1: .../data_file.txt') format'text' (delimiter ';' null '\N' escap...
-                                                             ^
-HINT:  Use the escape string syntax for escapes, e.g., E'\r\n'.
-WARNING:  nonstandard use of \' in a string literal
-LINE 1: ...xt') format'text' (delimiter ';' null '\N' escape '\') encod...
-                                                             ^
-HINT:  Use '' to write quotes in strings, or use the escape string syntax (E'...').
-2021-01-11 14:36:40|INFO|gpload session started 2021-01-11 14:36:40
-2021-01-11 14:36:40|INFO|setting schema 'public' for table 'testspecialchar'
-2021-01-11 14:36:40|INFO|started gpfdist -p 8081 -P 8082 -f "/home/gpadmin/workspace/gpdb/gpMgmt/bin/gpload_test/gpload2/data_file.txt" -t 30
-2021-01-11 14:36:40|INFO|did not find an external table to reuse. creating ext_gpload_reusable_5cfb15b6_53d7_11eb_ae6a_00505698707d
-2021-01-11 14:36:40|ERROR|unexpected error -- backtrace written to log file
-2021-01-11 14:36:40|INFO|rows Inserted          = 0
-2021-01-11 14:36:40|INFO|rows Updated           = 0
-2021-01-11 14:36:40|INFO|data formatting errors = 0
-2021-01-11 14:36:40|INFO|gpload failed
-NOTICE:  Table doesn't have 'DISTRIBUTED BY' clause -- Using column named 'Field1' as the Greenplum Database data distribution key for this table.
-HINT:  The 'DISTRIBUTED BY' clause determines the distribution of data. Make sure column(s) chosen are the optimal data distribution key to minimize skew.
-WARNING:  nonstandard use of \\ in a string literal
-LINE 13:                  and pgext.fmtopts like '%delimiter '';'' nu...
-                                                 ^
-HINT:  Use the escape string syntax for backslashes, e.g., E'\\'.
-WARNING:  nonstandard use of escape in a string literal
-LINE 1: .../data_file.txt') format'text' (delimiter ';' null '\N' escap...
-                                                             ^
-HINT:  Use the escape string syntax for escapes, e.g., E'\r\n'.
-WARNING:  nonstandard use of \' in a string literal
-LINE 1: ...xt') format'text' (delimiter ';' null '\N' escape '\') encod...
-                                                             ^
-HINT:  Use '' to write quotes in strings, or use the escape string syntax (E'...').
-2021-01-11 14:36:41|INFO|gpload session started 2021-01-11 14:36:41
-2021-01-11 14:36:41|INFO|setting schema 'public' for table 'testspecialchar'
-2021-01-11 14:36:41|INFO|started gpfdist -p 8081 -P 8082 -f "/home/gpadmin/workspace/gpdb/gpMgmt/bin/gpload_test/gpload2/data_file.txt" -t 30
-2021-01-11 14:36:41|INFO|did not find a staging table to reuse. creating staging_gpload_reusable_40df9a45044f2d17b97f89bbbc58f24f
-2021-01-11 14:36:41|INFO|did not find an external table to reuse. creating ext_gpload_reusable_5d335b1a_53d7_11eb_b06b_00505698707d
-2021-01-11 14:36:41|ERROR|unexpected error -- backtrace written to log file
-2021-01-11 14:36:41|INFO|rows Inserted          = 0
-2021-01-11 14:36:41|INFO|rows Updated           = 0
-2021-01-11 14:36:41|INFO|data formatting errors = 0
-2021-01-11 14:36:41|INFO|gpload failed
-2021-01-11 14:36:41|INFO|gpload session started 2021-01-11 14:36:41
-2021-01-11 14:36:41|INFO|setting schema 'public' for table 'testspecialchar'
-2021-01-11 14:36:41|ERROR|no mapping for input column "'Field1'" to output table
-2021-01-11 14:36:41|INFO|rows Inserted          = 0
-2021-01-11 14:36:41|INFO|rows Updated           = 0
-2021-01-11 14:36:41|INFO|data formatting errors = 0
-2021-01-11 14:36:41|INFO|gpload failed
-2021-01-11 14:36:41|INFO|gpload session started 2021-01-11 14:36:41
-2021-01-11 14:36:41|INFO|setting schema 'public' for table 'testspecialchar'
-2021-01-11 14:36:41|ERROR|no mapping for input column "Field1" to output table
-2021-01-11 14:36:41|INFO|rows Inserted          = 0
-2021-01-11 14:36:41|INFO|rows Updated           = 0
-2021-01-11 14:36:41|INFO|data formatting errors = 0
-2021-01-11 14:36:41|INFO|gpload failed
