+++ gpMgmt/bin/gpload_test/gpload2/setup.ans
@@ -13,13 +13,10 @@ DROP TABLE IF EXISTS texttable;
 DROP TABLE
 DROP TABLE IF EXISTS csvtable;
 DROP TABLE
-DROP TABLE IF EXISTS test.csvtable;
-DROP TABLE
 DROP TABLE IF EXISTS texttable1;
 DROP TABLE
 DROP TABLE IF EXISTS testSpecialChar;
-DROP TABLE
-DROP TABLE IF EXISTS chinese表;
+NOTICE:  table "testspecialchar" does not exist, skipping
 DROP TABLE
 reset client_min_messages;
 RESET
@@ -42,6 +39,4 @@ CREATE TABLE test.csvtable (
             DISTRIBUTED BY (year);
 CREATE TABLE
 create table testSpecialChar("Field1" bigint, "Field#2" text) distributed by ("Field1");
-CREATE TABLE
-CREATE TABLE  chinese表 ( 列1 text, "列#2" int, lie3 timestamp, 列four decimal ) DISTRIBUTED BY ("列#2");
 CREATE TABLE
