+++ gpMgmt/bin/gpload_test/gpload2/setup.sql
@@ -10,10 +10,8 @@ set client_min_messages='warning';
 DROP EXTERNAL TABLE IF EXISTS temp_gpload_staging_table;
 DROP TABLE IF EXISTS texttable;
 DROP TABLE IF EXISTS csvtable;
-DROP TABLE IF EXISTS test.csvtable;
 DROP TABLE IF EXISTS texttable1;
 DROP TABLE IF EXISTS testSpecialChar;
-DROP TABLE IF EXISTS chinese表;
 reset client_min_messages;
 CREATE TABLE texttable (
             s1 text, s2 text, s3 text, dt timestamp,
@@ -30,4 +28,3 @@ CREATE TABLE test.csvtable (
 	    year int, make text, model text, decription text, price decimal)
             DISTRIBUTED BY (year);
 create table testSpecialChar("Field1" bigint, "Field#2" text) distributed by ("Field1");
-CREATE TABLE  chinese表 ( 列1 text, "列#2" int, lie3 timestamp, 列four decimal ) DISTRIBUTED BY ("列#2");
